requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.6
django-htmlmin==0.11.0
django-storages==1.14.4
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gunicorn==23.0.0
html5lib==1.1
idna==3.8
minify_html==0.11.1
mysqlclient==2.2.4
openpyxl==3.1.4
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.7.0
